requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
python-telegram-bot-21.4
requests==2.31.0
